[jaeger] Do Not Decode Response (#590)
We do not decode the response from Jaeger anymore, because kobs consumes
a lot of memory for large traces. Instead we just read the body and
directly write it.

Note: This could be improved in the future to directly write the
response body from Jaeger, but then we would have to rework the error
handling, so that we decided to try this quick solution first.