[runbooks] Improve Sync
We are now only adding runbooks in the sync process when they are not
already present in another cluster, so that changes which are made in
one cluster for testing not overwritten by the same runbook in another
cluster.

We also automatically updating the runbooks in the frontend when a user
clicks on the sync button, so that the current version is shown and not
an old one after the sync.