[velero] Add Velero Plugin (#601)
Add a Velero plugin which can be used to view backups, restores and
schedules for Velero. The plugin can also be used to create new backups.
For this a template can be specified via the
"frontendOptions.backupTemplate" paramter. This template is shown to the
user and can be adjusted before the backup is created.