[velero] Create Backups from Schedules
It is now possible to create a Velero backup from a schedule, similar to
how it can be done via the `--from-schedule` flag in the Velero CLI.

For that a new button was added to the Velero schedules resources table
(`PlayArrow` icon), which triggers an API call to the new `/backup` API
endpoint. This endpoint implements a similar logic as the Velero CLI to
create the backup from the selected schedule.